Basic Changes
It was made two changes, one in the index.html file and the other in the
styles.css file. The first change was a text addition in the <p> tag,
telling new users to make more changes. The second, was a font change,
from Courier New to Times New Roman, coder style option.