Added an alt text for that image and expressed my fealings for this great project in the html body.
It is good practice to add an alt text to image tags in markup for the sake of the completeness
and strictness that is nowadays so absent on the world wide web. Therefore I took some time
and thinking about it, then worked hard at night until this solution was completed.

I hope this will be accepted into the the master branch.